chore: moved ui to their respective folders
