+++ bot/ui/modals/announcement.py
@@ -1,14 +1,14 @@
 from discord import Interaction, TextStyle, Attachment, TextChannel
 from discord.ui import Modal, TextInput
 
-from ui.views import AnnouncementView
+from ui.views.announcement import AnnouncementView
 
 
 class Announcement(Modal, title='Announcement'):
     announcement_title = TextInput(
         label='Title',
         placeholder='Title',
-        required=True,
+        required=False,
     )
     announcement = TextInput(
         label='Announcement',
@@ -47,7 +47,7 @@ class Announcement(Modal, title='Announcement'):
             channel_mentions = [channel.mention for channel in selection_view.channel_mentions]
 
             for tag, mention_type in zip(tags, (user_mentions, role_mentions, channel_mentions)):
-                if len(mention_type) > 0:  # If the list is empty (any of the following mention lists)
+                if len(mention_type) > 0:  # If the list is not empty (any of the following mention lists)
                     for mention in mention_type:
                         # Replace the tags one by one
                         announcement = announcement.replace(tag, mention, 1)
